+++ apheleia-utils.el
@@ -48,6 +48,7 @@ always returns nil to defer to the formatter."
               (js2-mode 'js2-basic-offset)
               (js3-mode 'js3-indent-level)
               (json-mode 'js-indent-level)
+              (js-json-mode 'js-indent-level)
               (json-ts-mode 'json-ts-mode-indent-offset)
               (nxml-mode 'nxml-child-indent)
               (robot-mode 'robot-mode-basic-offset)
@@ -61,7 +62,8 @@ always returns nil to defer to the formatter."
               (tsx-ts-mode 'typescript-ts-mode-indent-offset)
               (typescript-mode 'typescript-indent-level)
               (typescript-ts-mode 'typescript-ts-mode-indent-offset)
-              (yaml-mode 'yaml-indent-offset))))
+              (yaml-mode 'yaml-indent-offset)
+              (yaml-ts-mode 'yaml-indent-offset))))
 
     (when-let ((indent (and indent-var
                             (boundp indent-var)