+++ test/formatters/samplecode/gdformat/in.gd
@@ -1,0 +1,9 @@
+class X:
+	var x=[1,2,{'a':1}]
+	var y=[1,2,3,]     # trailing comma
+	func foo(a:int,b,c=[1,2,3]):
+		if a in c and \
+		   b > 100:
+			print('foo')
+func bar():
+	print('bar')