+++ test/formatters/samplecode/oxfmt/in.js
@@ -1,0 +1,4 @@
+function HelloWorld({greeting = "hello", greeted = '"World"', silent = false, onMouseOver,}) {
+
+  if(!greeting){return null};
+  }